feat: new DefiPlaza strategy
This new strategy looks at `balanceOf` and the `quoteRewards` function of the DFP2 governance contract to be able to use the balance of unclaimed staking rewards in the voting process, next to the balance of DFP2 in a wallet.

The addresses in the `examples.json` are the current biggest DFP2 holders to create a real-world test scenario.