Update index.ts
Would it be possible to implement a change similar to this idea? (perhaps as a separate strategy?)
Where instead of only having the option of having a range, it will have the tokenId and weight? 
Perhaps a range or tokenID built-in as an option? { length: (end - start) || tokenId / step + 1) an so on?