Refactoring to be easier to follow.
This also catches an edge case where an account is delegating to someone else, has a balance, and also has votes because someone else is delegating to them. The votes returned from getVotes are always counted no matter what, just like in the erc20-votes strategy. The local token balance is only added when an account is delegating to different valid address.