[erc20-votes-with-override] Adding ERC-20 votes with override strategy. (#448)
* Adding ERC-20 votes with override strategy.

* Fixing typo.

* Removing unnecessary toLowerCase call and external references.

* Make sure to map before filtering so the index doesn't get messed up.

* Changing Object types to Record

* Refactoring to be easier to follow.

This also catches an edge case where an account is delegating to someone else, has a balance, and also has votes because someone else is delegating to them. The votes returned from getVotes are always counted no matter what, just like in the erc20-votes strategy. The local token balance is only added when an account is delegating to different valid address.