[erc20-votes-with-override] Adding compatibility for Snapshot delegations. (#463)
* Adding compatibility for Snapshot delegations.

The standard "delegation" strategy is not compatible with the overridden voting power calculation, because it can lead to double-counting.

So instead, this strategy adds an "includeSnapshotDelegations" option. When enabled, one additional request is made to the subgraph to get Snapshot delegations. Then the unique delegators are added to the list of addresses and the normal overridden voting power calculation is used as before.

The "isSnapshotDelegatedScore" option determines whether the delegated or non-delegated score is used. So two strategies can be used, one with this option enabled and one with this option disabled.

* Adding more information and examples into readme.

* Fixing formatting issues in readme.

* Fixing formatting issues in readme.