Add WINRProtocol Governance Voting Strategy [winr-staking] (#1179)
* feat(winr-staking): add new strategy for Winr governance staking
This commit adds a new strategy for Winr governance staking. The strategy allows users that staked WINR and vWINR to vote on proposals by delegating their voting power to a delegatee. The strategy calculates the total staked amount for each user by summing up the staked amount from each active index, dividend stakes, and vested dividend stakes. The strategy is implemented in the `src/strategies/winr-staking` directory and includes a README.md file, examples.json file, index.ts file, and schema.json file.

* refactor(winr-staking): replace Promise.all with map to improve performance
The Promise.all function is replaced with map to improve performance. The multicaller.call function is now called for each address in the addresses array. This change allows multicaller to make the calls concurrently, which can improve performance.

* refactor(winr-staking): remove unnecessary Promise.all() and await from addresses.map()
The Promise.all() and await are not necessary in the addresses.map() function as the multicaller2.call() function returns a Promise. Removing the Promise.all() and await will not affect the functionality of the code and will make it more concise.

---------

Co-authored-by: Chaitanya <yourchaitu@gmail.com>